fix chinese character display issue this needed the str passed to Record in marc_tools.py to be encoded as utf-8 also typos/formatting replace VMs zoom.py with the original one, better documented
