+++ z3950/Marc/marc8_mapping.py
@@ -6,337 +6,6 @@
 
 """MARC-8 mapping."""
 
-CHARSET_34 = {  # Extended Arabic
-    0xA1: (0x6FD, 0),  # DOUBLE ALEF WITH HAMZA ABOVE / ARABIC SIGN SINDHI AMPERSAND
-    0xA2: (0x672, 0),  # ARABIC LETTER ALEF WITH WAVY HAMZA ABOVE
-    0xA3: (0x673, 0),  # ARABIC LETTER ALEF WITH WAVY HAMZA BELOW
-    0xA4: (0x679, 0),  # ARABIC LETTER TTEH
-    0xA5: (0x67A, 0),  # ARABIC LETTER TTEHEH
-    0xA6: (0x67B, 0),  # ARABIC LETTER BBEH
-    0xA7: (0x67C, 0),  # ARABIC LETTER TEH WITH RING
-    0xA8: (0x67D, 0),  # ARABIC LETTER TEH WITH THREE DOTS ABOVE DOWNWARDS
-    0xA9: (0x67E, 0),  # ARABIC LETTER PEH
-    0xAA: (0x67F, 0),  # ARABIC LETTER TEHEH
-    0xAB: (0x680, 0),  # ARABIC LETTER BEHEH
-    0xAC: (0x681, 0),  # ARABIC LETTER HAH WITH HAMZA ABOVE
-    0xAD: (0x682, 0),  # ARABIC LETTER HAH WITH TWO ABOVE DOTS VERTICAL ABOVE
-    0xAE: (0x683, 0),  # ARABIC LETTER NYEH
-    0xAF: (0x684, 0),  # ARABIC LETTER DYEH
-    0xB0: (0x685, 0),  # ARABIC LETTER HAH WITH THREE DOTS ABOVE
-    0xB1: (0x686, 0),  # ARABIC LETTER TCHEH
-    0xB2: (0x6BF, 0),  # ARABIC LETTER TCHEH WITH DOT ABOVE
-    0xB3: (0x687, 0),  # ARABIC LETTER TCHEHEH
-    0xB4: (0x688, 0),  # ARABIC LETTER DDAL
-    0xB5: (0x689, 0),  # ARABIC LETTER DAL WITH RING
-    0xB6: (0x68A, 0),  # ARABIC LETTER DAL WITH DOT BELOW
-    0xB7: (0x68B, 0),  # ARABIC LETTER DAL WITH DOT BELOW AND SMALL TAH
-    0xB8: (0x68C, 0),  # ARABIC LETTER DAHAL
-    0xB9: (0x68D, 0),  # ARABIC LETTER DDAHAL
-    0xBA: (0x68E, 0),  # ARABIC LETTER DUL
-    0xBB: (0x68F, 0),  # ARABIC LETTER DAL WITH THREE DOTS ABOVE DOWNWARDS
-    0xBC: (0x690, 0),  # ARABIC LETTER DAL WITH FOUR DOTS ABOVE
-    0xBD: (0x691, 0),  # ARABIC LETTER RREH
-    0xBE: (0x692, 0),  # ARABIC LETTER REH WITH SMALL V
-    0xBF: (0x693, 0),  # ARABIC LETTER REH WITH RING
-    0xC0: (0x694, 0),  # ARABIC LETTER REH WITH DOT BELOW
-    0xC1: (0x695, 0),  # ARABIC LETTER REH WITH SMALL V BELOW
-    0xC2: (0x696, 0),  # ARABIC LETTER REH WITH DOT BELOW AND DOT ABOVE
-    0xC3: (0x697, 0),  # ARABIC LETTER REH WITH TWO DOTS ABOVE
-    0xC4: (0x698, 0),  # ARABIC LETTER JEH
-    0xC5: (0x699, 0),  # ARABIC LETTER REH WITH FOUR DOTS ABOVE
-    0xC6: (0x69A, 0),  # ARABIC LETTER SEEN WITH DOT BELOW AND DOT ABOVE
-    0xC7: (0x69B, 0),  # ARABIC LETTER SEEN WITH THREE DOTS BELOW
-    0xC8: (0x69C, 0),  # ARABIC LETTER SEEN WITH THREE DOTS BELOW AND THREE DOTS ABOVE
-    0xC9: (0x6FA, 0),  # ARABIC LETTER SHEEN WITH DOT BELOW
-    0xCA: (0x69D, 0),  # ARABIC LETTER SAD WITH TWO DOTS BELOW
-    0xCB: (0x69E, 0),  # ARABIC LETTER SAD WITH THREE DOTS ABOVE
-    0xCC: (0x6FB, 0),  # ARABIC LETTER DAD WITH DOT BELOW
-    0xCD: (0x69F, 0),  # ARABIC LETTER TAH WITH THREE DOTS ABOVE
-    0xCE: (0x6A0, 0),  # ARABIC LETTER AIN WITH THREE DOTS ABOVE
-    0xCF: (0x6FC, 0),  # ARABIC LETTER GHAIN WITH DOT BELOW
-    0xD0: (0x6A1, 0),  # ARABIC LETTER DOTLESS FEH
-    0xD1: (0x6A2, 0),  # ARABIC LETTER FEH WITH DOT MOVED BELOW
-    0xD2: (0x6A3, 0),  # ARABIC LETTER FEH WITH DOT BELOW
-    0xD3: (0x6A4, 0),  # ARABIC LETTER VEH
-    0xD4: (0x6A5, 0),  # ARABIC LETTER FEH WITH THREE DOTS BELOW
-    0xD5: (0x6A6, 0),  # ARABIC LETTER PEHEH
-    0xD6: (0x6A7, 0),  # ARABIC LETTER QAF WITH DOT ABOVE
-    0xD7: (0x6A8, 0),  # ARABIC LETTER QAF WITH THREE DOTS ABOVE
-    0xD8: (0x6A9, 0),  # ARABIC LETTER KEHEH
-    0xD9: (0x6AA, 0),  # ARABIC LETTER SWASH KAF
-    0xDA: (0x6AB, 0),  # ARABIC LETTER KAF WITH RING
-    0xDB: (0x6AC, 0),  # ARABIC LETTER KAF WITH DOT ABOVE
-    0xDC: (0x6AD, 0),  # ARABIC LETTER NG
-    0xDD: (0x6AE, 0),  # ARABIC LETTER KAF WITH THREE DOTS BELOW
-    0xDE: (0x6AF, 0),  # ARABIC LETTER GAF
-    0xDF: (0x6B0, 0),  # ARABIC LETTER GAF WITH RING
-    0xE0: (0x6B1, 0),  # ARABIC LETTER NGOEH
-    0xE1: (0x6B2, 0),  # ARABIC LETTER GAF WITH TWO DOTS BELOW
-    0xE2: (0x6B3, 0),  # ARABIC LETTER GUEH
-    0xE3: (0x6B4, 0),  # ARABIC LETTER GAF WITH THREE DOTS ABOVE
-    0xE4: (0x6B5, 0),  # ARABIC LETTER LAM WITH SMALL V
-    0xE5: (0x6B6, 0),  # ARABIC LETTER LAM WITH DOT ABOVE
-    0xE6: (0x6B7, 0),  # ARABIC LETTER LAM WITH THREE DOTS ABOVE
-    0xE7: (0x6B8, 0),  # ARABIC LETTER LAM WITH THREE DOTS BELOW
-    0xE8: (0x6BA, 0),  # ARABIC LETTER NOON GHUNNA
-    0xE9: (0x6BB, 0),  # ARABIC LETTER RNOON
-    0xEA: (0x6BC, 0),  # ARABIC LETTER NOON WITH RING
-    0xEB: (0x6BD, 0),  # ARABIC LETTER NOON WITH THREE DOTS ABOVE
-    0xEC: (0x6B9, 0),  # ARABIC LETTER NOON WITH DOT BELOW
-    0xED: (0x6BE, 0),  # ARABIC LETTER HEH DOACHASHMEE
-    0xEE: (0x6C0, 0),  # HEH WITH HAMZA ABOVE / ARABIC LETTER HEH WITH YEH ABOVE
-    0xEF: (0x6C4, 0),  # ARABIC LETTER WAW WITH RING
-    0xF0: (0x6C5, 0),  # KYRGHYZ OE / ARABIC LETTER KIRGHIZ OE
-    0xF1: (0x6C6, 0),  # ARABIC LETTER OE
-    0xF2: (0x6CA, 0),  # ARABIC LETTER WAW WITH TWO DOTS ABOVE
-    0xF3: (0x6CB, 0),  # ARABIC LETTER VE
-    0xF4: (0x6CD, 0),  # ARABIC LETTER YEH WITH TAIL
-    0xF5: (0x6CE, 0),  # ARABIC LETTER YEH WITH SMALL V
-    0xF6: (0x6D0, 0),  # ARABIC LETTER E
-    0xF7: (0x6D2, 0),  # ARABIC LETTER YEH BARREE
-    0xF8: (0x6D3, 0),  # ARABIC LETTER YEH BARREE WITH HAMZA ABOVE
-    0xFD: (0x306, 1),  # SHORT E / COMBINING BREVE
-    0xFE: (0x30C, 1),  # SHORT U / COMBINING CARON
-}
-CHARSET_45 = {  # Extended Latin (ANSEL)
-    0x88: (0x98, 0),  # NON-SORT BEGIN / START OF STRING
-    0x89: (0x9C, 0),  # NON-SORT END / STRING TERMINATOR
-    0x8D: (0x200D, 0),  # JOINER / ZERO WIDTH JOINER
-    0x8E: (0x200C, 0),  # NON-JOINER / ZERO WIDTH NON-JOINER
-    0xA1: (0x141, 0),  # UPPERCASE POLISH L / LATIN CAPITAL LETTER L WITH STROKE
-    0xA2: (0xD8, 0),  # UPPERCASE SCANDINAVIAN O / LATIN CAPITAL LETTER O WITH STROKE
-    0xA3: (0x110, 0),  # UPPERCASE D WITH CROSSBAR / LATIN CAPITAL LETTER D WITH STROKE
-    0xA4: (
-        0xDE,
-        0,
-    ),  # UPPERCASE ICELANDIC THORN / LATIN CAPITAL LETTER THORN (Icelandic)
-    0xA5: (0xC6, 0),  # UPPERCASE DIGRAPH AE / LATIN CAPITAL LIGATURE AE
-    0xA6: (0x152, 0),  # UPPERCASE DIGRAPH OE / LATIN CAPITAL LIGATURE OE
-    0xA7: (0x2B9, 0),  # SOFT SIGN, PRIME / MODIFIER LETTER PRIME
-    0xA8: (0xB7, 0),  # MIDDLE DOT
-    0xA9: (0x266D, 0),  # MUSIC FLAT SIGN
-    0xAA: (0xAE, 0),  # PATENT MARK / REGISTERED SIGN
-    0xAB: (0xB1, 0),  # PLUS OR MINUS / PLUS-MINUS SIGN
-    0xAC: (0x1A0, 0),  # UPPERCASE O-HOOK / LATIN CAPITAL LETTER O WITH HORN
-    0xAD: (0x1AF, 0),  # UPPERCASE U-HOOK / LATIN CAPITAL LETTER U WITH HORN
-    0xAE: (0x2BC, 0),  # ALIF / MODIFIER LETTER RIGHT HALF RING
-    0xB0: (0x2BB, 0),  # AYN / MODIFIER LETTER TURNED COMMA
-    0xB1: (0x142, 0),  # LOWERCASE POLISH L / LATIN SMALL LETTER L WITH STROKE
-    0xB2: (0xF8, 0),  # LOWERCASE SCANDINAVIAN O / LATIN SMALL LETTER O WITH STROKE
-    0xB3: (0x111, 0),  # LOWERCASE D WITH CROSSBAR / LATIN SMALL LETTER D WITH STROKE
-    0xB4: (0xFE, 0),  # LOWERCASE ICELANDIC THORN / LATIN SMALL LETTER THORN (Icelandic)
-    0xB5: (0xE6, 0),  # LOWERCASE DIGRAPH AE / LATIN SMALL LIGATURE AE
-    0xB6: (0x153, 0),  # LOWERCASE DIGRAPH OE / LATIN SMALL LIGATURE OE
-    0xB7: (0x2BA, 0),  # HARD SIGN, DOUBLE PRIME / MODIFIER LETTER DOUBLE PRIME
-    0xB8: (0x131, 0),  # LOWERCASE TURKISH I / LATIN SMALL LETTER DOTLESS I
-    0xB9: (0xA3, 0),  # BRITISH POUND / POUND SIGN
-    0xBA: (0xF0, 0),  # LOWERCASE ETH / LATIN SMALL LETTER ETH (Icelandic)
-    0xBC: (0x1A1, 0),  # LOWERCASE O-HOOK / LATIN SMALL LETTER O WITH HORN
-    0xBD: (0x1B0, 0),  # LOWERCASE U-HOOK / LATIN SMALL LETTER U WITH HORN
-    0xC0: (0xB0, 0),  # DEGREE SIGN
-    0xC1: (0x2113, 0),  # SCRIPT SMALL L
-    0xC2: (0x2117, 0),  # SOUND RECORDING COPYRIGHT
-    0xC3: (0xA9, 0),  # COPYRIGHT SIGN
-    0xC4: (0x266F, 0),  # MUSIC SHARP SIGN
-    0xC5: (0xBF, 0),  # INVERTED QUESTION MARK
-    0xC6: (0xA1, 0),  # INVERTED EXCLAMATION MARK
-    0xC7: (0xDF, 0),  # ESZETT SYMBOL
-    0xC8: (0x20AC, 0),  # EURO SIGN
-    0xE0: (0x309, 1),  # PSEUDO QUESTION MARK / COMBINING HOOK ABOVE
-    0xE1: (0x300, 1),  # GRAVE / COMBINING GRAVE ACCENT (Varia)
-    0xE2: (0x301, 1),  # ACUTE / COMBINING ACUTE ACCENT (Oxia)
-    0xE3: (0x302, 1),  # CIRCUMFLEX / COMBINING CIRCUMFLEX ACCENT
-    0xE4: (0x303, 1),  # TILDE / COMBINING TILDE
-    0xE5: (0x304, 1),  # MACRON / COMBINING MACRON
-    0xE6: (0x306, 1),  # BREVE / COMBINING BREVE (Vrachy)
-    0xE7: (0x307, 1),  # SUPERIOR DOT / COMBINING DOT ABOVE
-    0xE8: (0x308, 1),  # UMLAUT, DIAERESIS / COMBINING DIAERESIS (Dialytika)
-    0xE9: (0x30C, 1),  # HACEK / COMBINING CARON
-    0xEA: (0x30A, 1),  # CIRCLE ABOVE, ANGSTROM / COMBINING RING ABOVE
-    0xEB: (0xFE20, 1),  # LIGATURE, FIRST HALF / COMBINING LIGATURE LEFT HALF
-    0xEC: (0xFE21, 1),  # LIGATURE, SECOND HALF / COMBINING LIGATURE RIGHT HALF
-    0xED: (0x315, 1),  # HIGH COMMA, OFF CENTER / COMBINING COMMA ABOVE RIGHT
-    0xEE: (0x30B, 1),  # DOUBLE ACUTE / COMBINING DOUBLE ACUTE ACCENT
-    0xEF: (0x310, 1),  # CANDRABINDU / COMBINING CANDRABINDU
-    0xF0: (0x327, 1),  # CEDILLA / COMBINING CEDILLA
-    0xF1: (0x328, 1),  # RIGHT HOOK, OGONEK / COMBINING OGONEK
-    0xF2: (0x323, 1),  # DOT BELOW / COMBINING DOT BELOW
-    0xF3: (0x324, 1),  # DOUBLE DOT BELOW / COMBINING DIAERESIS BELOW
-    0xF4: (0x325, 1),  # CIRCLE BELOW / COMBINING RING BELOW
-    0xF5: (0x333, 1),  # DOUBLE UNDERSCORE / COMBINING DOUBLE LOW LINE
-    0xF6: (0x332, 1),  # UNDERSCORE / COMBINING LOW LINE
-    0xF7: (0x326, 1),  # LEFT HOOK (COMMA BELOW) / COMBINING COMMA BELOW
-    0xF8: (0x31C, 1),  # RIGHT CEDILLA / COMBINING LEFT HALF RING BELOW
-    0xF9: (0x32E, 1),  # UPADHMANIYA / COMBINING BREVE BELOW
-    0xFA: (0xFE22, 1),  # DOUBLE TILDE, FIRST HALF / COMBINING DOUBLE TILDE LEFT HALF
-    0xFB: (0xFE23, 1),  # DOUBLE TILDE, SECOND HALF / COMBINING DOUBLE TILDE RIGHT HALF
-    0xFE: (0x313, 1),  # HIGH COMMA, CENTERED / COMBINING COMMA ABOVE (Psili)
-}
-CHARSET_33 = {  # Basic Arabic
-    0x21: (0x21, 0),  # EXCLAMATION MARK
-    0x22: (0x22, 0),  # QUOTATION MARK
-    0x23: (0x23, 0),  # NUMBER SIGN
-    0x24: (0x24, 0),  # DOLLAR SIGN
-    0x25: (0x66A, 0),  # PERCENT SIGN / ARABIC PERCENT SIGN
-    0x26: (0x26, 0),  # AMPERSAND
-    0x27: (0x27, 0),  # APOSTROPHE
-    0x28: (0x28, 0),  # OPENING PARENTHESIS / LEFT PARENTHESIS
-    0x29: (0x29, 0),  # CLOSING PARENTHESIS / RIGHT PARENTHESIS
-    0x2A: (0x66D, 0),  # ASTERISK / ARABIC FIVE POINTED STAR
-    0x2B: (0x2B, 0),  # PLUS SIGN
-    0x2C: (0x60C, 0),  # ARABIC COMMA
-    0x2D: (0x2D, 0),  # HYPHEN-MINUS
-    0x2E: (0x2E, 0),  # PERIOD, DECIMAL POINT / FULL STOP
-    0x2F: (0x2F, 0),  # SLASH / SOLIDUS
-    0x30: (0x660, 0),  # ARABIC-INDIC DIGIT ZERO
-    0x31: (0x661, 0),  # ARABIC-INDIC DIGIT ONE
-    0x32: (0x662, 0),  # ARABIC-INDIC DIGIT TWO
-    0x33: (0x663, 0),  # ARABIC-INDIC DIGIT THREE
-    0x34: (0x664, 0),  # ARABIC-INDIC DIGIT FOUR
-    0x35: (0x665, 0),  # ARABIC-INDIC DIGIT FIVE
-    0x36: (0x666, 0),  # ARABIC-INDIC DIGIT SIX
-    0x37: (0x667, 0),  # ARABIC-INDIC DIGIT SEVEN
-    0x38: (0x668, 0),  # ARABIC-INDIC DIGIT EIGHT
-    0x39: (0x669, 0),  # ARABIC-INDIC DIGIT NINE
-    0x3A: (0x3A, 0),  # COLON
-    0x3B: (0x61B, 0),  # ARABIC SEMICOLON
-    0x3C: (0x3C, 0),  # LESS-THAN SIGN
-    0x3D: (0x3D, 0),  # EQUALS SIGN
-    0x3E: (0x3E, 0),  # GREATER-THAN SIGN
-    0x3F: (0x61F, 0),  # ARABIC QUESTION MARK
-    0x41: (0x621, 0),  # HAMZAH / ARABIC LETTER HAMZA
-    0x42: (0x622, 0),  # ARABIC LETTER ALEF WITH MADDA ABOVE
-    0x43: (0x623, 0),  # ARABIC LETTER ALEF WITH HAMZA ABOVE
-    0x44: (0x624, 0),  # ARABIC LETTER WAW WITH HAMZA ABOVE
-    0x45: (0x625, 0),  # ARABIC LETTER ALEF WITH HAMZA BELOW
-    0x46: (0x626, 0),  # ARABIC LETTER YEH WITH HAMZA ABOVE
-    0x47: (0x627, 0),  # ARABIC LETTER ALEF
-    0x48: (0x628, 0),  # ARABIC LETTER BEH
-    0x49: (0x629, 0),  # ARABIC LETTER TEH MARBUTA
-    0x4A: (0x62A, 0),  # ARABIC LETTER TEH
-    0x4B: (0x62B, 0),  # ARABIC LETTER THEH
-    0x4C: (0x62C, 0),  # ARABIC LETTER JEEM
-    0x4D: (0x62D, 0),  # ARABIC LETTER HAH
-    0x4E: (0x62E, 0),  # ARABIC LETTER KHAH
-    0x4F: (0x62F, 0),  # ARABIC LETTER DAL
-    0x50: (0x630, 0),  # ARABIC LETTER THAL
-    0x51: (0x631, 0),  # ARABIC LETTER REH
-    0x52: (0x632, 0),  # ARABIC LETTER ZAIN
-    0x53: (0x633, 0),  # ARABIC LETTER SEEN
-    0x54: (0x634, 0),  # ARABIC LETTER SHEEN
-    0x55: (0x635, 0),  # ARABIC LETTER SAD
-    0x56: (0x636, 0),  # ARABIC LETTER DAD
-    0x57: (0x637, 0),  # ARABIC LETTER TAH
-    0x58: (0x638, 0),  # ARABIC LETTER ZAH
-    0x59: (0x639, 0),  # ARABIC LETTER AIN
-    0x5A: (0x63A, 0),  # ARABIC LETTER GHAIN
-    0x5B: (0x5B, 0),  # OPENING SQUARE BRACKET / LEFT SQUARE BRACKET
-    0x5D: (0x5D, 0),  # CLOSING SQUARE BRACKET / RIGHT SQUARE BRACKET
-    0x60: (0x640, 0),  # ARABIC TATWEEL
-    0x61: (0x641, 0),  # ARABIC LETTER FEH
-    0x62: (0x642, 0),  # ARABIC LETTER QAF
-    0x63: (0x643, 0),  # ARABIC LETTER KAF
-    0x64: (0x644, 0),  # ARABIC LETTER LAM
-    0x65: (0x645, 0),  # ARABIC LETTER MEEM
-    0x66: (0x646, 0),  # ARABIC LETTER NOON
-    0x67: (0x647, 0),  # ARABIC LETTER HEH
-    0x68: (0x648, 0),  # ARABIC LETTER WAW
-    0x69: (0x649, 0),  # ARABIC LETTER ALEF MAKSURA
-    0x6A: (0x64A, 0),  # ARABIC LETTER YEH
-    0x6B: (0x64B, 1),  # ARABIC FATHATAN
-    0x6C: (0x64C, 1),  # ARABIC DAMMATAN
-    0x6D: (0x64D, 1),  # ARABIC KASRATAN
-    0x6E: (0x64E, 1),  # ARABIC FATHA
-    0x6F: (0x64F, 1),  # ARABIC DAMMA
-    0x70: (0x650, 1),  # ARABIC KASRA
-    0x71: (0x651, 1),  # ARABIC SHADDA
-    0x72: (0x652, 1),  # ARABIC SUKUN
-    0x73: (0x671, 0),  # ARABIC LETTER ALEF WASLA
-    0x74: (0x670, 0),  # ARABIC LETTER SUPERSCRIPT ALEF
-    0x78: (0x66C, 0),  # ARABIC THOUSANDS SEPARATOR
-    0x79: (0x201D, 0),  # RIGHT DOUBLE QUOTATION MARK
-    0x7A: (0x201C, 0),  # LEFT DOUBLE QUOTATION MARK
-}
-CHARSET_32 = {  # Basic Hebrew
-    0x21: (0x21, 0),  # EXCLAMATION MARK
-    0x22: (0x5F4, 0),  # QUOTATION MARK, GERSHAYIM / HEBREW PUNCTUATION GERSHAYIM
-    0x23: (0x23, 0),  # NUMBER SIGN
-    0x24: (0x24, 0),  # DOLLAR SIGN
-    0x25: (0x25, 0),  # PERCENT SIGN
-    0x26: (0x26, 0),  # AMPERSAND
-    0x27: (0x5F3, 0),  # APOSTROPHE, GERESH / HEBREW PUNCTUATION GERESH
-    0x28: (0x28, 0),  # OPENING PARENTHESIS / LEFT PARENTHESIS
-    0x29: (0x29, 0),  # CLOSING PARENTHESIS / RIGHT PARENTHESIS
-    0x2A: (0x2A, 0),  # ASTERISK
-    0x2B: (0x2B, 0),  # PLUS SIGN
-    0x2C: (0x2C, 0),  # COMMA
-    0x2D: (0x5BE, 0),  # HYPHEN-MINUS, MAKEF / HEBREW PUNCTUATION MAQAF
-    0x2E: (0x2E, 0),  # PERIOD, DECIMAL POINT / FULL STOP
-    0x2F: (0x2F, 0),  # SLASH / SOLIDUS
-    0x30: (0x30, 0),  # DIGIT ZERO
-    0x31: (0x31, 0),  # DIGIT ONE
-    0x32: (0x32, 0),  # DIGIT TWO
-    0x33: (0x33, 0),  # DIGIT THREE
-    0x34: (0x34, 0),  # DIGIT FOUR
-    0x35: (0x35, 0),  # DIGIT FIVE
-    0x36: (0x36, 0),  # DIGIT SIX
-    0x37: (0x37, 0),  # DIGIT SEVEN
-    0x38: (0x38, 0),  # DIGIT EIGHT
-    0x39: (0x39, 0),  # DIGIT NINE
-    0x3A: (0x3A, 0),  # COLON
-    0x3B: (0x3B, 0),  # SEMICOLON
-    0x3C: (0x3C, 0),  # LESS-THAN SIGN
-    0x3D: (0x3D, 0),  # EQUALS SIGN
-    0x3E: (0x3E, 0),  # GREATER-THAN SIGN
-    0x3F: (0x3F, 0),  # QUESTION MARK
-    0x40: (0x5B7, 1),  # HEBREW POINT PATAH
-    0x41: (0x5B8, 1),  # KAMATS / HEBREW POINT QAMATS
-    0x42: (0x5B6, 1),  # HEBREW POINT SEGOL
-    0x43: (0x5B5, 1),  # TSEREH / HEBREW POINT TSERE
-    0x44: (0x5B4, 1),  # HIRIK / HEBREW POINT HIRIQ
-    0x45: (0x5B9, 1),  # HOLAM, LEFT SIN DOT / HEBREW POINT HOLAM
-    0x46: (0x5BB, 1),  # KUBUTS / HEBREW POINT QUBUTS
-    0x47: (0x5B0, 1),  # HEBREW POINT SHEVA
-    0x48: (0x5B2, 1),  # HEBREW POINT HATAF PATAH
-    0x49: (0x5B3, 1),  # HATAF KAMATS / HEBREW POINT HATAF QAMATS
-    0x4A: (0x5B1, 1),  # HEBREW POINT HATAF SEGOL
-    0x4B: (0x5BC, 1),  # HEBREW POINT DAGESH OR MAPIQ
-    0x4C: (0x5BF, 1),  # RAFEH / HEBREW POINT RAFE
-    0x4D: (0x5C1, 1),  # RIGHT SHIN DOT / HEBREW POINT  SHIN DOT
-    0x4E: (0xFB1E, 1),  # VARIKA / HEBREW POINT JUDEO-SPANISH VARIKA
-    0x5B: (0x5B, 0),  # OPENING SQUARE BRACKET / LEFT SQUARE BRACKET
-    0x5D: (0x5D, 0),  # CLOSING SQUARE BRACKET / RIGHT SQUARE BRACKET
-    0x60: (0x5D0, 0),  # HEBREW LETTER ALEF
-    0x61: (0x5D1, 0),  # HEBREW LETTER BET
-    0x62: (0x5D2, 0),  # HEBREW LETTER GIMEL
-    0x63: (0x5D3, 0),  # HEBREW LETTER DALET
-    0x64: (0x5D4, 0),  # HEBREW LETTER HE
-    0x65: (0x5D5, 0),  # HEBREW LETTER VAV
-    0x66: (0x5D6, 0),  # HEBREW LETTER ZAYIN
-    0x67: (0x5D7, 0),  # HEBREW LETTER HET
-    0x68: (0x5D8, 0),  # HEBREW LETTER TET
-    0x69: (0x5D9, 0),  # HEBREW LETTER YOD
-    0x6A: (0x5DA, 0),  # HEBREW LETTER FINAL KAF
-    0x6B: (0x5DB, 0),  # HEBREW LETTER KAF
-    0x6C: (0x5DC, 0),  # HEBREW LETTER LAMED
-    0x6D: (0x5DD, 0),  # HEBREW LETTER FINAL MEM
-    0x6E: (0x5DE, 0),  # HEBREW LETTER MEM
-    0x6F: (0x5DF, 0),  # HEBREW LETTER FINAL NUN
-    0x70: (0x5E0, 0),  # HEBREW LETTER NUN
-    0x71: (0x5E1, 0),  # HEBREW LETTER SAMEKH
-    0x72: (0x5E2, 0),  # HEBREW LETTER AYIN
-    0x73: (0x5E3, 0),  # HEBREW LETTER FINAL PE
-    0x74: (0x5E4, 0),  # HEBREW LETTER PE
-    0x75: (0x5E5, 0),  # HEBREW LETTER FINAL TSADI
-    0x76: (0x5E6, 0),  # HEBREW LETTER TSADI
-    0x77: (0x5E7, 0),  # HEBREW LETTER QOF / KOF
-    0x78: (0x5E8, 0),  # HEBREW LETTER RESH
-    0x79: (0x5E9, 0),  # HEBREW LETTER SHIN
-    0x7A: (0x5EA, 0),  # HEBREW LETTER TAV
-    0x7B: (0x5F0, 0),  # HEBREW LIGATURE YIDDISH DOUBLE VAV / TSVEY VOVN
-    0x7C: (0x5F1, 0),  # HEBREW LIGATURE YIDDISH VAV YOD / VOV YUD
-    0x7D: (0x5F2, 0),  # HEBREW LIGATURE YIDDISH DOUBLE YOD / TSVEY YUDN
-}
 CHARSET_31 = {  # Chinese, Japanese, Korean (EACC)
     0x215556: (0x8461, 0),  # East Asian ideograph
     0x6F5557: (0xC5D1, 0),  # Korean hangul
@@ -419,10 +88,7 @@ CHARSET_31 = {  # Chinese, Japanese, Korean (EACC)
     0x6F5147: (0xBCF6, 0),  # Korean hangul
     0x6F557B: (0xC639, 0),  # Korean hangul
     0x215945: (0x8B66, 0),  # East Asian ideograph
-    0x21372C: (
-        0x5606,
-        0,
-    ),  # East Asian ideograph (variant of 4B372C which maps to 5606)
+    0x21372C: (0x5606, 0),  # East Asian ideograph (variant of 4B372C which maps to 5606)
     0x27557C: (0x829C, 0),  # East Asian ideograph
     0x224027: (0x69BF, 0),  # East Asian ideograph
     0x23557D: (0x9B34, 0),  # East Asian ideograph
@@ -449,10 +115,7 @@ CHARSET_31 = {  # Chinese, Japanese, Korean (EACC)
     0x6F514A: (0xBD04, 0),  # Korean hangul
     0x335F73: (0x9759, 0),  # East Asian ideograph
     0x6F4C33: (0xB17C, 0),  # Korean hangul
-    0x4B525B: (
-        0x66DC,
-        0,
-    ),  # East Asian ideograph (variant of 39525B which maps to 66DC)
+    0x4B525B: (0x66DC, 0),  # East Asian ideograph (variant of 39525B which maps to 66DC)
     0x2E7C2E: (0x831C, 0),  # East Asian ideograph
     0x224038: (0x69C3, 0),  # East Asian ideograph
     0x6F5B4D: (0xD280, 0),  # Korean hangul
@@ -559,10 +222,7 @@ CHARSET_31 = {  # Chinese, Japanese, Korean (EACC)
     0x23406D: (0x9185, 0),  # East Asian ideograph
     0x22406E: (0x6A18, 0),  # East Asian ideograph
     0x23406F: (0x9182, 0),  # East Asian ideograph
-    0x4B5F30: (
-        0x9686,
-        0,
-    ),  # East Asian ideograph (variant of 215F30 which maps to 9686)
+    0x4B5F30: (0x9686, 0),  # East Asian ideograph (variant of 215F30 which maps to 9686)
     0x234070: (0x918A, 0),  # East Asian ideograph
     0x6F5A75: (0xD0D0, 0),  # Korean hangul
     0x213C38: (0x5DE8, 0),  # East Asian ideograph
@@ -602,10 +262,7 @@ CHARSET_31 = {  # Chinese, Japanese, Korean (EACC)
     0x6F503B: (0xBAA9, 0),  # Korean hangul
     0x222B2D: (0x602B, 0),  # East Asian ideograph
     0x6F5B50: (0xD290, 0),  # Korean hangul
-    0x455847: (
-        0x8A25,
-        0,
-    ),  # East Asian ideograph (variant of 215847 which maps to 8A25)
+    0x455847: (0x8A25, 0),  # East Asian ideograph (variant of 215847 which maps to 8A25)
     0x396B2F: (0x521F, 0),  # East Asian ideograph
     0x6F515A: (0xBD84, 0),  # Korean hangul
     0x6F5861: (0xCAD9, 0),  # Korean hangul
@@ -613,10 +270,7 @@ CHARSET_31 = {  # Chinese, Japanese, Korean (EACC)
     0x225D57: (0x754E, 0),  # East Asian ideograph
     0x4B3052: (0x6275, 0),  # East Asian ideograph
     0x212B31: (0xFF3B, 0),  # Ideographic left square bracket
-    0x4B3749: (
-        0x5668,
-        0,
-    ),  # East Asian ideograph (variant of 213749 which maps to 5668)
+    0x4B3749: (0x5668, 0),  # East Asian ideograph (variant of 213749 which maps to 5668)
     0x3A3B7D: (0x67B1, 0),  # East Asian ideograph
     0x216B33: (0x5231, 0),  # East Asian ideograph
     0x23504A: (0x98BF, 0),  # East Asian ideograph
@@ -696,14 +350,8 @@ CHARSET_31 = {  # Chinese, Japanese, Korean (EACC)
     0x345452: (0x7118, 0),  # East Asian ideograph
     0x6F5B40: (0xD1D8, 0),  # Korean hangul
     0x213569: (0x5462, 0),  # East Asian ideograph
-    0x4C6B62: (
-        0x7B4C,
-        0,
-    ),  # East Asian ideograph (variant of 226B62 which maps to 7B4C)
-    0x394634: (
-        0x6B96,
-        0,
-    ),  # East Asian ideograph (variant of 214634 which maps to 6B96)
+    0x4C6B62: (0x7B4C, 0),  # East Asian ideograph (variant of 226B62 which maps to 7B4C)
+    0x394634: (0x6B96, 0),  # East Asian ideograph (variant of 214634 which maps to 6B96)
     0x274171: (0x629A, 0),  # East Asian ideograph
     0x6F5165: (0xBDF0, 0),  # Korean hangul
     0x28356D: (0x6512, 0),  # East Asian ideograph
@@ -712,10 +360,7 @@ CHARSET_31 = {  # Chinese, Japanese, Korean (EACC)
     0x4B3D27: (0x5EC3, 0),  # East Asian ideograph
     0x6F5166: (0xBE0C, 0),  # Korean hangul
     0x6F4D23: (0xB310, 0),  # Korean hangul
-    0x213B61: (
-        0x5C64,
-        0,
-    ),  # East Asian ideograph (variant of 4B3B61 which maps to 5C64)
+    0x213B61: (0x5C64, 0),  # East Asian ideograph (variant of 4B3B61 which maps to 5C64)
     0x4B5277: (0x8068, 0),  # East Asian ideograph
     0x336162: (0x9A23, 0),  # East Asian ideograph
     0x705F51: (0x54D0, 0),  # East Asian ideograph
@@ -817,10 +462,7 @@ CHARSET_31 = {  # Chinese, Japanese, Korean (EACC)
     0x225D6D: (0x7579, 0),  # East Asian ideograph
     0x21564A: (0x865F, 0),  # East Asian ideograph
     0x2D563C: (0x8613, 0),  # East Asian ideograph
-    0x45564B: (
-        0x865E,
-        0,
-    ),  # East Asian ideograph (variant of 21564B which maps to 865E)
+    0x45564B: (0x865E, 0),  # East Asian ideograph (variant of 21564B which maps to 865E)
     0x21564C: (0x8667, 0),  # East Asian ideograph
     0x6F5171: (0xBE4C, 0),  # Korean hangul
     0x27605C: (0x98D2, 0),  # East Asian ideograph
@@ -891,10 +533,7 @@ CHARSET_31 = {  # Chinese, Japanese, Korean (EACC)
     0x22403D: (0x6A12, 0),  # East Asian ideograph
     0x6F5669: (0xC758, 0),  # Korean hangul
     0x224C67: (0x6F34, 0),  # East Asian ideograph
-    0x4B566A: (
-        0x8708,
-        0,
-    ),  # East Asian ideograph (variant of 21566A which maps to 8708)
+    0x4B566A: (0x8708, 0),  # East Asian ideograph (variant of 21566A which maps to 8708)
     0x21566B: (0x8700, 0),  # East Asian ideograph
     0x276062: (0x98D8, 0),  # East Asian ideograph
     0x285E7A: (0x75D6, 0),  # East Asian ideograph
@@ -16495,156 +16134,267 @@ CHARSET_31 = {  # Chinese, Japanese, Korean (EACC)
     ),  # East Asian ideograph (variant of 213F22 which maps to 6168)
     0x275555: (0x83B4, 0),  # East Asian ideograph
 }
-CHARSET_70 = {  # Superscripts
-    0x28: (0x207D, 0),  # SUPERSCRIPT OPENING PARENTHESIS / SUPERSCRIPT LEFT PARENTHESIS
-    0x29: (
-        0x207E,
-        0,
-    ),  # SUPERSCRIPT CLOSING PARENTHESIS / SUPERSCRIPT RIGHT PARENTHESIS
-    0x2B: (0x207A, 0),  # SUPERSCRIPT PLUS SIGN
-    0x2D: (0x207B, 0),  # SUPERSCRIPT HYPHEN-MINUS / SUPERSCRIPT MINUS
-    0x30: (0x2070, 0),  # SUPERSCRIPT DIGIT ZERO
-    0x31: (0xB9, 0),  # SUPERSCRIPT DIGIT ONE
-    0x32: (0xB2, 0),  # SUPERSCRIPT DIGIT TWO
-    0x33: (0xB3, 0),  # SUPERSCRIPT DIGIT THREE
-    0x34: (0x2074, 0),  # SUPERSCRIPT DIGIT FOUR
-    0x35: (0x2075, 0),  # SUPERSCRIPT DIGIT FIVE
-    0x36: (0x2076, 0),  # SUPERSCRIPT DIGIT SIX
-    0x37: (0x2077, 0),  # SUPERSCRIPT DIGIT SEVEN
-    0x38: (0x2078, 0),  # SUPERSCRIPT DIGIT EIGHT
-    0x39: (0x2079, 0),  # SUPERSCRIPT DIGIT NINE
+
+CHARSET_32 = {  # Basic Hebrew
+    0x21: (0x21, 0),  # EXCLAMATION MARK
+    0x22: (0x5F4, 0),  # QUOTATION MARK, GERSHAYIM / HEBREW PUNCTUATION GERSHAYIM
+    0x23: (0x23, 0),  # NUMBER SIGN
+    0x24: (0x24, 0),  # DOLLAR SIGN
+    0x25: (0x25, 0),  # PERCENT SIGN
+    0x26: (0x26, 0),  # AMPERSAND
+    0x27: (0x5F3, 0),  # APOSTROPHE, GERESH / HEBREW PUNCTUATION GERESH
+    0x28: (0x28, 0),  # OPENING PARENTHESIS / LEFT PARENTHESIS
+    0x29: (0x29, 0),  # CLOSING PARENTHESIS / RIGHT PARENTHESIS
+    0x2A: (0x2A, 0),  # ASTERISK
+    0x2B: (0x2B, 0),  # PLUS SIGN
+    0x2C: (0x2C, 0),  # COMMA
+    0x2D: (0x5BE, 0),  # HYPHEN-MINUS, MAKEF / HEBREW PUNCTUATION MAQAF
+    0x2E: (0x2E, 0),  # PERIOD, DECIMAL POINT / FULL STOP
+    0x2F: (0x2F, 0),  # SLASH / SOLIDUS
+    0x30: (0x30, 0),  # DIGIT ZERO
+    0x31: (0x31, 0),  # DIGIT ONE
+    0x32: (0x32, 0),  # DIGIT TWO
+    0x33: (0x33, 0),  # DIGIT THREE
+    0x34: (0x34, 0),  # DIGIT FOUR
+    0x35: (0x35, 0),  # DIGIT FIVE
+    0x36: (0x36, 0),  # DIGIT SIX
+    0x37: (0x37, 0),  # DIGIT SEVEN
+    0x38: (0x38, 0),  # DIGIT EIGHT
+    0x39: (0x39, 0),  # DIGIT NINE
+    0x3A: (0x3A, 0),  # COLON
+    0x3B: (0x3B, 0),  # SEMICOLON
+    0x3C: (0x3C, 0),  # LESS-THAN SIGN
+    0x3D: (0x3D, 0),  # EQUALS SIGN
+    0x3E: (0x3E, 0),  # GREATER-THAN SIGN
+    0x3F: (0x3F, 0),  # QUESTION MARK
+    0x40: (0x5B7, 1),  # HEBREW POINT PATAH
+    0x41: (0x5B8, 1),  # KAMATS / HEBREW POINT QAMATS
+    0x42: (0x5B6, 1),  # HEBREW POINT SEGOL
+    0x43: (0x5B5, 1),  # TSEREH / HEBREW POINT TSERE
+    0x44: (0x5B4, 1),  # HIRIK / HEBREW POINT HIRIQ
+    0x45: (0x5B9, 1),  # HOLAM, LEFT SIN DOT / HEBREW POINT HOLAM
+    0x46: (0x5BB, 1),  # KUBUTS / HEBREW POINT QUBUTS
+    0x47: (0x5B0, 1),  # HEBREW POINT SHEVA
+    0x48: (0x5B2, 1),  # HEBREW POINT HATAF PATAH
+    0x49: (0x5B3, 1),  # HATAF KAMATS / HEBREW POINT HATAF QAMATS
+    0x4A: (0x5B1, 1),  # HEBREW POINT HATAF SEGOL
+    0x4B: (0x5BC, 1),  # HEBREW POINT DAGESH OR MAPIQ
+    0x4C: (0x5BF, 1),  # RAFEH / HEBREW POINT RAFE
+    0x4D: (0x5C1, 1),  # RIGHT SHIN DOT / HEBREW POINT  SHIN DOT
+    0x4E: (0xFB1E, 1),  # VARIKA / HEBREW POINT JUDEO-SPANISH VARIKA
+    0x5B: (0x5B, 0),  # OPENING SQUARE BRACKET / LEFT SQUARE BRACKET
+    0x5D: (0x5D, 0),  # CLOSING SQUARE BRACKET / RIGHT SQUARE BRACKET
+    0x60: (0x5D0, 0),  # HEBREW LETTER ALEF
+    0x61: (0x5D1, 0),  # HEBREW LETTER BET
+    0x62: (0x5D2, 0),  # HEBREW LETTER GIMEL
+    0x63: (0x5D3, 0),  # HEBREW LETTER DALET
+    0x64: (0x5D4, 0),  # HEBREW LETTER HE
+    0x65: (0x5D5, 0),  # HEBREW LETTER VAV
+    0x66: (0x5D6, 0),  # HEBREW LETTER ZAYIN
+    0x67: (0x5D7, 0),  # HEBREW LETTER HET
+    0x68: (0x5D8, 0),  # HEBREW LETTER TET
+    0x69: (0x5D9, 0),  # HEBREW LETTER YOD
+    0x6A: (0x5DA, 0),  # HEBREW LETTER FINAL KAF
+    0x6B: (0x5DB, 0),  # HEBREW LETTER KAF
+    0x6C: (0x5DC, 0),  # HEBREW LETTER LAMED
+    0x6D: (0x5DD, 0),  # HEBREW LETTER FINAL MEM
+    0x6E: (0x5DE, 0),  # HEBREW LETTER MEM
+    0x6F: (0x5DF, 0),  # HEBREW LETTER FINAL NUN
+    0x70: (0x5E0, 0),  # HEBREW LETTER NUN
+    0x71: (0x5E1, 0),  # HEBREW LETTER SAMEKH
+    0x72: (0x5E2, 0),  # HEBREW LETTER AYIN
+    0x73: (0x5E3, 0),  # HEBREW LETTER FINAL PE
+    0x74: (0x5E4, 0),  # HEBREW LETTER PE
+    0x75: (0x5E5, 0),  # HEBREW LETTER FINAL TSADI
+    0x76: (0x5E6, 0),  # HEBREW LETTER TSADI
+    0x77: (0x5E7, 0),  # HEBREW LETTER QOF / KOF
+    0x78: (0x5E8, 0),  # HEBREW LETTER RESH
+    0x79: (0x5E9, 0),  # HEBREW LETTER SHIN
+    0x7A: (0x5EA, 0),  # HEBREW LETTER TAV
+    0x7B: (0x5F0, 0),  # HEBREW LIGATURE YIDDISH DOUBLE VAV / TSVEY VOVN
+    0x7C: (0x5F1, 0),  # HEBREW LIGATURE YIDDISH VAV YOD / VOV YUD
+    0x7D: (0x5F2, 0),  # HEBREW LIGATURE YIDDISH DOUBLE YOD / TSVEY YUDN
 }
-CHARSET_51 = {  # Extended Cyrillic
-    0xC0: (
-        0x491,
-        0,
-    ),  # LOWERCASE GE WITH UPTURN / CYRILLIC SMALL LETTER GHE WITH UPTURN
-    0xC1: (0x452, 0),  # LOWERCASE DJE / CYRILLIC SMALL LETTER DJE (Serbian)
-    0xC2: (0x453, 0),  # CYRILLIC SMALL LETTER GJE
-    0xC3: (0x454, 0),  # LOWERCASE E / CYRILLIC SMALL LETTER UKRAINIAN IE
-    0xC4: (0x451, 0),  # CYRILLIC SMALL LETTER IO
-    0xC5: (0x455, 0),  # CYRILLIC SMALL LETTER DZE
-    0xC6: (0x456, 0),  # LOWERCASE I / CYRILLIC SMALL LETTER BYELORUSSIAN-UKRANIAN I
-    0xC7: (0x457, 0),  # LOWERCASE YI / CYRILLIC SMALL LETTER YI (Ukrainian)
-    0xC8: (0x458, 0),  # CYRILLIC SMALL LETTER JE
-    0xC9: (0x459, 0),  # CYRILLIC SMALL LETTER LJE
-    0xCA: (0x45A, 0),  # CYRILLIC SMALL LETTER NJE
-    0xCB: (0x45B, 0),  # LOWERCASE TSHE / CYRILLIC SMALL LETTER TSHE (Serbian)
-    0xCC: (0x45C, 0),  # CYRILLIC SMALL LETTER KJE
-    0xCD: (
-        0x45E,
-        0,
-    ),  # LOWERCASE SHORT U / CYRILLIC SMALL LETTER SHORT U (Byelorussian)
-    0xCE: (0x45F, 0),  # CYRILLIC SMALL LETTER DZHE
-    0xD0: (0x463, 0),  # CYRILLIC SMALL LETTER YAT
-    0xD1: (0x473, 0),  # CYRILLIC SMALL LETTER FITA
-    0xD2: (0x475, 0),  # CYRILLIC SMALL LETTER IZHITSA
-    0xD3: (0x46B, 0),  # CYRILLIC SMALL LETTER BIG YUS
-    0xDB: (0x5B, 0),  # OPENING SQUARE BRACKET / LEFT SQUARE BRACKET
-    0xDD: (0x5D, 0),  # CLOSING SQUARE BRACKET / RIGHT SQUARE BRACKET
-    0xDF: (0x5F, 0),  # SPACING UNDERSCORE / LOW LINE
-    0xE0: (
-        0x490,
-        0,
-    ),  # UPPERCASE GE WITH UPTURN / CYRILLIC CAPITAL LETTER GHE WITH UPTURN
-    0xE1: (0x402, 0),  # UPPERCASE DJE / CYRILLIC CAPITAL LETTER DJE (Serbian)
-    0xE2: (0x403, 0),  # CYRILLIC CAPITAL LETTER GJE
-    0xE3: (0x404, 0),  # UPPERCASE E / CYRILLIC CAPITAL LETTER UKRAINIAN IE
-    0xE4: (0x401, 0),  # CYRILLIC CAPITAL LETTER IO
-    0xE5: (0x405, 0),  # CYRILLIC CAPITAL LETTER DZE
-    0xE6: (0x406, 0),  # UPPERCASE I / CYRILLIC CAPITAL LETTER BYELORUSSIAN-UKRANIAN I
-    0xE7: (0x407, 0),  # UPPERCASE YI / CYRILLIC CAPITAL LETTER YI (Ukrainian)
-    0xE8: (0x408, 0),  # CYRILLIC CAPITAL LETTER JE
-    0xE9: (0x409, 0),  # CYRILLIC CAPITAL LETTER LJE
-    0xEA: (0x40A, 0),  # CYRILLIC CAPITAL LETTER NJE
-    0xEB: (0x40B, 0),  # UPPERCASE TSHE / CYRILLIC CAPITAL LETTER TSHE (Serbian)
-    0xEC: (0x40C, 0),  # CYRILLIC CAPITAL LETTER KJE
-    0xED: (
-        0x40E,
-        0,
-    ),  # UPPERCASE SHORT U / CYRILLIC CAPITAL LETTER SHORT U (Byelorussian)
-    0xEE: (0x40F, 0),  # CYRILLIC CAPITAL LETTER DZHE
-    0xEF: (0x42A, 0),  # CYRILLIC CAPITAL LETTER HARD SIGN
-    0xF0: (0x462, 0),  # CYRILLIC CAPITAL LETTER YAT
-    0xF1: (0x472, 0),  # CYRILLIC CAPITAL LETTER FITA
-    0xF2: (0x474, 0),  # CYRILLIC CAPITAL LETTER IZHITSA
-    0xF3: (0x46A, 0),  # CYRILLIC CAPITAL LETTER BIG YUS
+
+CHARSET_33 = {  # Basic Arabic
+    0x21: (0x21, 0),  # EXCLAMATION MARK
+    0x22: (0x22, 0),  # QUOTATION MARK
+    0x23: (0x23, 0),  # NUMBER SIGN
+    0x24: (0x24, 0),  # DOLLAR SIGN
+    0x25: (0x66A, 0),  # PERCENT SIGN / ARABIC PERCENT SIGN
+    0x26: (0x26, 0),  # AMPERSAND
+    0x27: (0x27, 0),  # APOSTROPHE
+    0x28: (0x28, 0),  # OPENING PARENTHESIS / LEFT PARENTHESIS
+    0x29: (0x29, 0),  # CLOSING PARENTHESIS / RIGHT PARENTHESIS
+    0x2A: (0x66D, 0),  # ASTERISK / ARABIC FIVE POINTED STAR
+    0x2B: (0x2B, 0),  # PLUS SIGN
+    0x2C: (0x60C, 0),  # ARABIC COMMA
+    0x2D: (0x2D, 0),  # HYPHEN-MINUS
+    0x2E: (0x2E, 0),  # PERIOD, DECIMAL POINT / FULL STOP
+    0x2F: (0x2F, 0),  # SLASH / SOLIDUS
+    0x30: (0x660, 0),  # ARABIC-INDIC DIGIT ZERO
+    0x31: (0x661, 0),  # ARABIC-INDIC DIGIT ONE
+    0x32: (0x662, 0),  # ARABIC-INDIC DIGIT TWO
+    0x33: (0x663, 0),  # ARABIC-INDIC DIGIT THREE
+    0x34: (0x664, 0),  # ARABIC-INDIC DIGIT FOUR
+    0x35: (0x665, 0),  # ARABIC-INDIC DIGIT FIVE
+    0x36: (0x666, 0),  # ARABIC-INDIC DIGIT SIX
+    0x37: (0x667, 0),  # ARABIC-INDIC DIGIT SEVEN
+    0x38: (0x668, 0),  # ARABIC-INDIC DIGIT EIGHT
+    0x39: (0x669, 0),  # ARABIC-INDIC DIGIT NINE
+    0x3A: (0x3A, 0),  # COLON
+    0x3B: (0x61B, 0),  # ARABIC SEMICOLON
+    0x3C: (0x3C, 0),  # LESS-THAN SIGN
+    0x3D: (0x3D, 0),  # EQUALS SIGN
+    0x3E: (0x3E, 0),  # GREATER-THAN SIGN
+    0x3F: (0x61F, 0),  # ARABIC QUESTION MARK
+    0x41: (0x621, 0),  # HAMZAH / ARABIC LETTER HAMZA
+    0x42: (0x622, 0),  # ARABIC LETTER ALEF WITH MADDA ABOVE
+    0x43: (0x623, 0),  # ARABIC LETTER ALEF WITH HAMZA ABOVE
+    0x44: (0x624, 0),  # ARABIC LETTER WAW WITH HAMZA ABOVE
+    0x45: (0x625, 0),  # ARABIC LETTER ALEF WITH HAMZA BELOW
+    0x46: (0x626, 0),  # ARABIC LETTER YEH WITH HAMZA ABOVE
+    0x47: (0x627, 0),  # ARABIC LETTER ALEF
+    0x48: (0x628, 0),  # ARABIC LETTER BEH
+    0x49: (0x629, 0),  # ARABIC LETTER TEH MARBUTA
+    0x4A: (0x62A, 0),  # ARABIC LETTER TEH
+    0x4B: (0x62B, 0),  # ARABIC LETTER THEH
+    0x4C: (0x62C, 0),  # ARABIC LETTER JEEM
+    0x4D: (0x62D, 0),  # ARABIC LETTER HAH
+    0x4E: (0x62E, 0),  # ARABIC LETTER KHAH
+    0x4F: (0x62F, 0),  # ARABIC LETTER DAL
+    0x50: (0x630, 0),  # ARABIC LETTER THAL
+    0x51: (0x631, 0),  # ARABIC LETTER REH
+    0x52: (0x632, 0),  # ARABIC LETTER ZAIN
+    0x53: (0x633, 0),  # ARABIC LETTER SEEN
+    0x54: (0x634, 0),  # ARABIC LETTER SHEEN
+    0x55: (0x635, 0),  # ARABIC LETTER SAD
+    0x56: (0x636, 0),  # ARABIC LETTER DAD
+    0x57: (0x637, 0),  # ARABIC LETTER TAH
+    0x58: (0x638, 0),  # ARABIC LETTER ZAH
+    0x59: (0x639, 0),  # ARABIC LETTER AIN
+    0x5A: (0x63A, 0),  # ARABIC LETTER GHAIN
+    0x5B: (0x5B, 0),  # OPENING SQUARE BRACKET / LEFT SQUARE BRACKET
+    0x5D: (0x5D, 0),  # CLOSING SQUARE BRACKET / RIGHT SQUARE BRACKET
+    0x60: (0x640, 0),  # ARABIC TATWEEL
+    0x61: (0x641, 0),  # ARABIC LETTER FEH
+    0x62: (0x642, 0),  # ARABIC LETTER QAF
+    0x63: (0x643, 0),  # ARABIC LETTER KAF
+    0x64: (0x644, 0),  # ARABIC LETTER LAM
+    0x65: (0x645, 0),  # ARABIC LETTER MEEM
+    0x66: (0x646, 0),  # ARABIC LETTER NOON
+    0x67: (0x647, 0),  # ARABIC LETTER HEH
+    0x68: (0x648, 0),  # ARABIC LETTER WAW
+    0x69: (0x649, 0),  # ARABIC LETTER ALEF MAKSURA
+    0x6A: (0x64A, 0),  # ARABIC LETTER YEH
+    0x6B: (0x64B, 1),  # ARABIC FATHATAN
+    0x6C: (0x64C, 1),  # ARABIC DAMMATAN
+    0x6D: (0x64D, 1),  # ARABIC KASRATAN
+    0x6E: (0x64E, 1),  # ARABIC FATHA
+    0x6F: (0x64F, 1),  # ARABIC DAMMA
+    0x70: (0x650, 1),  # ARABIC KASRA
+    0x71: (0x651, 1),  # ARABIC SHADDA
+    0x72: (0x652, 1),  # ARABIC SUKUN
+    0x73: (0x671, 0),  # ARABIC LETTER ALEF WASLA
+    0x74: (0x670, 0),  # ARABIC LETTER SUPERSCRIPT ALEF
+    0x78: (0x66C, 0),  # ARABIC THOUSANDS SEPARATOR
+    0x79: (0x201D, 0),  # RIGHT DOUBLE QUOTATION MARK
+    0x7A: (0x201C, 0),  # LEFT DOUBLE QUOTATION MARK
 }
-CHARSET_53 = {  # Basic Greek
-    0x21: (0x300, 1),  # COMBINING GRAVE ACCENT
-    0x22: (0x301, 1),  # COMBINING ACUTE ACCENT
-    0x23: (0x308, 1),  # COMBINING DIAERESIS
-    0x24: (0x342, 1),  # COMBINING GREEK PERISPOMENI / CIRCUMFLEX
-    0x25: (0x313, 1),  # COMBINING COMMA ABOVE / SMOOTH BREATHING
-    0x26: (0x314, 1),  # COMBINING REVERSED COMMA ABOVE / ROUGH BREATHING
-    0x27: (0x345, 1),  # COMBINING GREEK YPOGEGRAMMENI / IOTA SUBSCRIPT
-    0x30: (0xAB, 0),  # LEFT-POINTING DOUBLE ANGLE QUOTATION MARK
-    0x31: (0xBB, 0),  # RIGHT-POINTING DOUBLE ANGLE QUOTATION MARK
-    0x32: (0x201C, 0),  # LEFT DOUBLE QUOTATION MARK
-    0x33: (0x201D, 0),  # RIGHT DOUBLE QUOTATION MARK
-    0x34: (0x374, 0),  # GREEK NUMERAL SIGN / UPPER PRIME
-    0x35: (0x375, 0),  # GREEK LOWER NUMERAL SIGN / LOWER PRIME
-    0x3B: (0x387, 0),  # GREEK ANO TELEIA / RAISED DOT, GREEK SEMICOLON
-    0x3F: (0x37E, 0),  # GREEK QUESTION MARK
-    0x41: (0x391, 0),  # GREEK CAPITAL LETTER ALPHA
-    0x42: (0x392, 0),  # GREEK CAPITAL LETTER BETA
-    0x44: (0x393, 0),  # GREEK CAPITAL LETTER GAMMA
-    0x45: (0x394, 0),  # GREEK CAPITAL LETTER DELTA
-    0x46: (0x395, 0),  # GREEK CAPITAL LETTER EPSILON
-    0x47: (0x3DA, 0),  # GREEK LETTER STIGMA
-    0x48: (0x3DC, 0),  # GREEK LETTER DIGAMMA
-    0x49: (0x396, 0),  # GREEK CAPITAL LETTER ZETA
-    0x4A: (0x397, 0),  # GREEK CAPITAL LETTER ETA
-    0x4B: (0x398, 0),  # GREEK CAPITAL LETTER THETA
-    0x4C: (0x399, 0),  # GREEK CAPITAL LETTER IOTA
-    0x4D: (0x39A, 0),  # GREEK CAPITAL LETTER KAPPA
-    0x4E: (0x39B, 0),  # GREEK CAPITAL LETTER LAMDA
-    0x4F: (0x39C, 0),  # GREEK CAPITAL LETTER MU
-    0x50: (0x39D, 0),  # GREEK CAPITAL LETTER NU
-    0x51: (0x39E, 0),  # GREEK CAPITAL LETTER XI
-    0x52: (0x39F, 0),  # GREEK CAPITAL LETTER OMICRON
-    0x53: (0x3A0, 0),  # GREEK CAPITAL LETTER PI
-    0x54: (0x3DE, 0),  # GREEK LETTER KOPPA
-    0x55: (0x3A1, 0),  # GREEK CAPITAL LETTER RHO
-    0x56: (0x3A3, 0),  # GREEK CAPITAL LETTER SIGMA
-    0x58: (0x3A4, 0),  # GREEK CAPITAL LETTER TAU
-    0x59: (0x3A5, 0),  # GREEK CAPITAL LETTER UPSILON
-    0x5A: (0x3A6, 0),  # GREEK CAPITAL LETTER PHI
-    0x5B: (0x3A7, 0),  # GREEK CAPITAL LETTER CHI
-    0x5C: (0x3A8, 0),  # GREEK CAPITAL LETTER PSI
-    0x5D: (0x3A9, 0),  # GREEK CAPITAL LETTER OMEGA
-    0x5E: (0x3E0, 0),  # GREEK LETTER SAMPI
-    0x61: (0x3B1, 0),  # GREEK SMALL LETTER ALPHA
-    0x62: (0x3B2, 0),  # GREEK SMALL LETTER BETA / SMALL LETTER BETA BEGINNING OF WORD
-    0x63: (0x3D0, 0),  # GREEK BETA SYMBOL / SMALL LETTER BETA MIDDLE OF WORD
-    0x64: (0x3B3, 0),  # GREEK SMALL LETTER GAMMA
-    0x65: (0x3B4, 0),  # GREEK SMALL LETTER DELTA
-    0x66: (0x3B5, 0),  # GREEK SMALL LETTER EPSILON
-    0x67: (0x3DB, 0),  # GREEK SMALL LETTER STIGMA
-    0x68: (0x3DD, 0),  # GREEK SMALL LETTER DIGAMMA
-    0x69: (0x3B6, 0),  # GREEK SMALL LETTER ZETA
-    0x6A: (0x3B7, 0),  # GREEK SMALL LETTER ETA
-    0x6B: (0x3B8, 0),  # GREEK SMALL LETTER THETA
-    0x6C: (0x3B9, 0),  # GREEK SMALL LETTER IOTA
-    0x6D: (0x3BA, 0),  # GREEK SMALL LETTER KAPPA
-    0x6E: (0x3BB, 0),  # GREEK SMALL LETTER LAMDA
-    0x6F: (0x3BC, 0),  # GREEK SMALL LETTER MU
-    0x70: (0x3BD, 0),  # GREEK SMALL LETTER NU
-    0x71: (0x3BE, 0),  # GREEK SMALL LETTER XI
-    0x72: (0x3BF, 0),  # GREEK SMALL LETTER OMICRON
-    0x73: (0x3C0, 0),  # GREEK SMALL LETTER PI
-    0x74: (0x3DF, 0),  # GREEK SMALL LETTER KOPPA
-    0x75: (0x3C1, 0),  # GREEK SMALL LETTER RHO
-    0x76: (0x3C3, 0),  # GREEK SMALL LETTER SIGMA
-    0x77: (0x3C2, 0),  # GREEK SMALL LETTER FINAL SIGMA / SMALL LETTER SIGMA END OF WORD
-    0x78: (0x3C4, 0),  # GREEK SMALL LETTER TAU
-    0x79: (0x3C5, 0),  # GREEK SMALL LETTER UPSILON
-    0x7A: (0x3C6, 0),  # GREEK SMALL LETTER PHI
-    0x7B: (0x3C7, 0),  # GREEK SMALL LETTER CHI
-    0x7C: (0x3C8, 0),  # GREEK SMALL LETTER PSI
-    0x7D: (0x3C9, 0),  # GREEK SMALL LETTER OMEGA
-    0x7E: (0x3E1, 0),  # GREEK SMALL LETTER SAMPI
+
+CHARSET_34 = {  # Extended Arabic
+    0xA1: (0x6FD, 0),  # DOUBLE ALEF WITH HAMZA ABOVE / ARABIC SIGN SINDHI AMPERSAND
+    0xA2: (0x672, 0),  # ARABIC LETTER ALEF WITH WAVY HAMZA ABOVE
+    0xA3: (0x673, 0),  # ARABIC LETTER ALEF WITH WAVY HAMZA BELOW
+    0xA4: (0x679, 0),  # ARABIC LETTER TTEH
+    0xA5: (0x67A, 0),  # ARABIC LETTER TTEHEH
+    0xA6: (0x67B, 0),  # ARABIC LETTER BBEH
+    0xA7: (0x67C, 0),  # ARABIC LETTER TEH WITH RING
+    0xA8: (0x67D, 0),  # ARABIC LETTER TEH WITH THREE DOTS ABOVE DOWNWARDS
+    0xA9: (0x67E, 0),  # ARABIC LETTER PEH
+    0xAA: (0x67F, 0),  # ARABIC LETTER TEHEH
+    0xAB: (0x680, 0),  # ARABIC LETTER BEHEH
+    0xAC: (0x681, 0),  # ARABIC LETTER HAH WITH HAMZA ABOVE
+    0xAD: (0x682, 0),  # ARABIC LETTER HAH WITH TWO ABOVE DOTS VERTICAL ABOVE
+    0xAE: (0x683, 0),  # ARABIC LETTER NYEH
+    0xAF: (0x684, 0),  # ARABIC LETTER DYEH
+    0xB0: (0x685, 0),  # ARABIC LETTER HAH WITH THREE DOTS ABOVE
+    0xB1: (0x686, 0),  # ARABIC LETTER TCHEH
+    0xB2: (0x6BF, 0),  # ARABIC LETTER TCHEH WITH DOT ABOVE
+    0xB3: (0x687, 0),  # ARABIC LETTER TCHEHEH
+    0xB4: (0x688, 0),  # ARABIC LETTER DDAL
+    0xB5: (0x689, 0),  # ARABIC LETTER DAL WITH RING
+    0xB6: (0x68A, 0),  # ARABIC LETTER DAL WITH DOT BELOW
+    0xB7: (0x68B, 0),  # ARABIC LETTER DAL WITH DOT BELOW AND SMALL TAH
+    0xB8: (0x68C, 0),  # ARABIC LETTER DAHAL
+    0xB9: (0x68D, 0),  # ARABIC LETTER DDAHAL
+    0xBA: (0x68E, 0),  # ARABIC LETTER DUL
+    0xBB: (0x68F, 0),  # ARABIC LETTER DAL WITH THREE DOTS ABOVE DOWNWARDS
+    0xBC: (0x690, 0),  # ARABIC LETTER DAL WITH FOUR DOTS ABOVE
+    0xBD: (0x691, 0),  # ARABIC LETTER RREH
+    0xBE: (0x692, 0),  # ARABIC LETTER REH WITH SMALL V
+    0xBF: (0x693, 0),  # ARABIC LETTER REH WITH RING
+    0xC0: (0x694, 0),  # ARABIC LETTER REH WITH DOT BELOW
+    0xC1: (0x695, 0),  # ARABIC LETTER REH WITH SMALL V BELOW
+    0xC2: (0x696, 0),  # ARABIC LETTER REH WITH DOT BELOW AND DOT ABOVE
+    0xC3: (0x697, 0),  # ARABIC LETTER REH WITH TWO DOTS ABOVE
+    0xC4: (0x698, 0),  # ARABIC LETTER JEH
+    0xC5: (0x699, 0),  # ARABIC LETTER REH WITH FOUR DOTS ABOVE
+    0xC6: (0x69A, 0),  # ARABIC LETTER SEEN WITH DOT BELOW AND DOT ABOVE
+    0xC7: (0x69B, 0),  # ARABIC LETTER SEEN WITH THREE DOTS BELOW
+    0xC8: (0x69C, 0),  # ARABIC LETTER SEEN WITH THREE DOTS BELOW AND THREE DOTS ABOVE
+    0xC9: (0x6FA, 0),  # ARABIC LETTER SHEEN WITH DOT BELOW
+    0xCA: (0x69D, 0),  # ARABIC LETTER SAD WITH TWO DOTS BELOW
+    0xCB: (0x69E, 0),  # ARABIC LETTER SAD WITH THREE DOTS ABOVE
+    0xCC: (0x6FB, 0),  # ARABIC LETTER DAD WITH DOT BELOW
+    0xCD: (0x69F, 0),  # ARABIC LETTER TAH WITH THREE DOTS ABOVE
+    0xCE: (0x6A0, 0),  # ARABIC LETTER AIN WITH THREE DOTS ABOVE
+    0xCF: (0x6FC, 0),  # ARABIC LETTER GHAIN WITH DOT BELOW
+    0xD0: (0x6A1, 0),  # ARABIC LETTER DOTLESS FEH
+    0xD1: (0x6A2, 0),  # ARABIC LETTER FEH WITH DOT MOVED BELOW
+    0xD2: (0x6A3, 0),  # ARABIC LETTER FEH WITH DOT BELOW
+    0xD3: (0x6A4, 0),  # ARABIC LETTER VEH
+    0xD4: (0x6A5, 0),  # ARABIC LETTER FEH WITH THREE DOTS BELOW
+    0xD5: (0x6A6, 0),  # ARABIC LETTER PEHEH
+    0xD6: (0x6A7, 0),  # ARABIC LETTER QAF WITH DOT ABOVE
+    0xD7: (0x6A8, 0),  # ARABIC LETTER QAF WITH THREE DOTS ABOVE
+    0xD8: (0x6A9, 0),  # ARABIC LETTER KEHEH
+    0xD9: (0x6AA, 0),  # ARABIC LETTER SWASH KAF
+    0xDA: (0x6AB, 0),  # ARABIC LETTER KAF WITH RING
+    0xDB: (0x6AC, 0),  # ARABIC LETTER KAF WITH DOT ABOVE
+    0xDC: (0x6AD, 0),  # ARABIC LETTER NG
+    0xDD: (0x6AE, 0),  # ARABIC LETTER KAF WITH THREE DOTS BELOW
+    0xDE: (0x6AF, 0),  # ARABIC LETTER GAF
+    0xDF: (0x6B0, 0),  # ARABIC LETTER GAF WITH RING
+    0xE0: (0x6B1, 0),  # ARABIC LETTER NGOEH
+    0xE1: (0x6B2, 0),  # ARABIC LETTER GAF WITH TWO DOTS BELOW
+    0xE2: (0x6B3, 0),  # ARABIC LETTER GUEH
+    0xE3: (0x6B4, 0),  # ARABIC LETTER GAF WITH THREE DOTS ABOVE
+    0xE4: (0x6B5, 0),  # ARABIC LETTER LAM WITH SMALL V
+    0xE5: (0x6B6, 0),  # ARABIC LETTER LAM WITH DOT ABOVE
+    0xE6: (0x6B7, 0),  # ARABIC LETTER LAM WITH THREE DOTS ABOVE
+    0xE7: (0x6B8, 0),  # ARABIC LETTER LAM WITH THREE DOTS BELOW
+    0xE8: (0x6BA, 0),  # ARABIC LETTER NOON GHUNNA
+    0xE9: (0x6BB, 0),  # ARABIC LETTER RNOON
+    0xEA: (0x6BC, 0),  # ARABIC LETTER NOON WITH RING
+    0xEB: (0x6BD, 0),  # ARABIC LETTER NOON WITH THREE DOTS ABOVE
+    0xEC: (0x6B9, 0),  # ARABIC LETTER NOON WITH DOT BELOW
+    0xED: (0x6BE, 0),  # ARABIC LETTER HEH DOACHASHMEE
+    0xEE: (0x6C0, 0),  # HEH WITH HAMZA ABOVE / ARABIC LETTER HEH WITH YEH ABOVE
+    0xEF: (0x6C4, 0),  # ARABIC LETTER WAW WITH RING
+    0xF0: (0x6C5, 0),  # KYRGHYZ OE / ARABIC LETTER KIRGHIZ OE
+    0xF1: (0x6C6, 0),  # ARABIC LETTER OE
+    0xF2: (0x6CA, 0),  # ARABIC LETTER WAW WITH TWO DOTS ABOVE
+    0xF3: (0x6CB, 0),  # ARABIC LETTER VE
+    0xF4: (0x6CD, 0),  # ARABIC LETTER YEH WITH TAIL
+    0xF5: (0x6CE, 0),  # ARABIC LETTER YEH WITH SMALL V
+    0xF6: (0x6D0, 0),  # ARABIC LETTER E
+    0xF7: (0x6D2, 0),  # ARABIC LETTER YEH BARREE
+    0xF8: (0x6D3, 0),  # ARABIC LETTER YEH BARREE WITH HAMZA ABOVE
+    0xFD: (0x306, 1),  # SHORT E / COMBINING BREVE
+    0xFE: (0x30C, 1),  # SHORT U / COMBINING CARON
 }
+
 CHARSET_42 = {  # Basic Latin (ASCII)
     0x1B: (0x1B, 0),  # ESCAPE (Unlikely to occur in UCS/Unicode)
     0x1D: (0x1D, 0),  # RECORD TERMINATOR / GROUP SEPARATOR
@@ -16746,27 +16496,79 @@ CHARSET_42 = {  # Basic Latin (ASCII)
     0x7D: (0x7D, 0),  # CLOSING CURLY BRACKET / RIGHT CURLY BRACKET
     0x7E: (0x7E, 0),  # SPACING TILDE / TILDE
 }
-CHARSET_62 = {  # Subscripts
-    0x28: (0x208D, 0),  # SUBSCRIPT OPENING PARENTHESIS / SUBSCRIPT LEFT PARENTHESIS
-    0x29: (0x208E, 0),  # SUBSCRIPT CLOSING PARENTHESIS / SUBSCRIPT RIGHT PARENTHESIS
-    0x2B: (0x208A, 0),  # SUBSCRIPT PLUS SIGN
-    0x2D: (0x208B, 0),  # SUBSCRIPT HYPHEN-MINUS / SUBSCRIPT MINUS
-    0x30: (0x2080, 0),  # SUBSCRIPT DIGIT ZERO
-    0x31: (0x2081, 0),  # SUBSCRIPT DIGIT ONE
-    0x32: (0x2082, 0),  # SUBSCRIPT DIGIT TWO
-    0x33: (0x2083, 0),  # SUBSCRIPT DIGIT THREE
-    0x34: (0x2084, 0),  # SUBSCRIPT DIGIT FOUR
-    0x35: (0x2085, 0),  # SUBSCRIPT DIGIT FIVE
-    0x36: (0x2086, 0),  # SUBSCRIPT DIGIT SIX
-    0x37: (0x2087, 0),  # SUBSCRIPT DIGIT SEVEN
-    0x38: (0x2088, 0),  # SUBSCRIPT DIGIT EIGHT
-    0x39: (0x2089, 0),  # SUBSCRIPT DIGIT NINE
+
+CHARSET_45 = {  # Extended Latin (ANSEL)
+    0x88: (0x98, 0),  # NON-SORT BEGIN / START OF STRING
+    0x89: (0x9C, 0),  # NON-SORT END / STRING TERMINATOR
+    0x8D: (0x200D, 0),  # JOINER / ZERO WIDTH JOINER
+    0x8E: (0x200C, 0),  # NON-JOINER / ZERO WIDTH NON-JOINER
+    0xA1: (0x141, 0),  # UPPERCASE POLISH L / LATIN CAPITAL LETTER L WITH STROKE
+    0xA2: (0xD8, 0),  # UPPERCASE SCANDINAVIAN O / LATIN CAPITAL LETTER O WITH STROKE
+    0xA3: (0x110, 0),  # UPPERCASE D WITH CROSSBAR / LATIN CAPITAL LETTER D WITH STROKE
+    0xA4: (0xDE, 0),  # UPPERCASE ICELANDIC THORN / LATIN CAPITAL LETTER THORN (Icelandic)
+    0xA5: (0xC6, 0),  # UPPERCASE DIGRAPH AE / LATIN CAPITAL LIGATURE AE
+    0xA6: (0x152, 0),  # UPPERCASE DIGRAPH OE / LATIN CAPITAL LIGATURE OE
+    0xA7: (0x2B9, 0),  # SOFT SIGN, PRIME / MODIFIER LETTER PRIME
+    0xA8: (0xB7, 0),  # MIDDLE DOT
+    0xA9: (0x266D, 0),  # MUSIC FLAT SIGN
+    0xAA: (0xAE, 0),  # PATENT MARK / REGISTERED SIGN
+    0xAB: (0xB1, 0),  # PLUS OR MINUS / PLUS-MINUS SIGN
+    0xAC: (0x1A0, 0),  # UPPERCASE O-HOOK / LATIN CAPITAL LETTER O WITH HORN
+    0xAD: (0x1AF, 0),  # UPPERCASE U-HOOK / LATIN CAPITAL LETTER U WITH HORN
+    0xAE: (0x2BC, 0),  # ALIF / MODIFIER LETTER RIGHT HALF RING
+    0xB0: (0x2BB, 0),  # AYN / MODIFIER LETTER TURNED COMMA
+    0xB1: (0x142, 0),  # LOWERCASE POLISH L / LATIN SMALL LETTER L WITH STROKE
+    0xB2: (0xF8, 0),  # LOWERCASE SCANDINAVIAN O / LATIN SMALL LETTER O WITH STROKE
+    0xB3: (0x111, 0),  # LOWERCASE D WITH CROSSBAR / LATIN SMALL LETTER D WITH STROKE
+    0xB4: (0xFE, 0),  # LOWERCASE ICELANDIC THORN / LATIN SMALL LETTER THORN (Icelandic)
+    0xB5: (0xE6, 0),  # LOWERCASE DIGRAPH AE / LATIN SMALL LIGATURE AE
+    0xB6: (0x153, 0),  # LOWERCASE DIGRAPH OE / LATIN SMALL LIGATURE OE
+    0xB7: (0x2BA, 0),  # HARD SIGN, DOUBLE PRIME / MODIFIER LETTER DOUBLE PRIME
+    0xB8: (0x131, 0),  # LOWERCASE TURKISH I / LATIN SMALL LETTER DOTLESS I
+    0xB9: (0xA3, 0),  # BRITISH POUND / POUND SIGN
+    0xBA: (0xF0, 0),  # LOWERCASE ETH / LATIN SMALL LETTER ETH (Icelandic)
+    0xBC: (0x1A1, 0),  # LOWERCASE O-HOOK / LATIN SMALL LETTER O WITH HORN
+    0xBD: (0x1B0, 0),  # LOWERCASE U-HOOK / LATIN SMALL LETTER U WITH HORN
+    0xC0: (0xB0, 0),  # DEGREE SIGN
+    0xC1: (0x2113, 0),  # SCRIPT SMALL L
+    0xC2: (0x2117, 0),  # SOUND RECORDING COPYRIGHT
+    0xC3: (0xA9, 0),  # COPYRIGHT SIGN
+    0xC4: (0x266F, 0),  # MUSIC SHARP SIGN
+    0xC5: (0xBF, 0),  # INVERTED QUESTION MARK
+    0xC6: (0xA1, 0),  # INVERTED EXCLAMATION MARK
+    0xC7: (0xDF, 0),  # ESZETT SYMBOL
+    0xC8: (0x20AC, 0),  # EURO SIGN
+    0xE0: (0x309, 1),  # PSEUDO QUESTION MARK / COMBINING HOOK ABOVE
+    0xE1: (0x300, 1),  # GRAVE / COMBINING GRAVE ACCENT (Varia)
+    0xE2: (0x301, 1),  # ACUTE / COMBINING ACUTE ACCENT (Oxia)
+    0xE3: (0x302, 1),  # CIRCUMFLEX / COMBINING CIRCUMFLEX ACCENT
+    0xE4: (0x303, 1),  # TILDE / COMBINING TILDE
+    0xE5: (0x304, 1),  # MACRON / COMBINING MACRON
+    0xE6: (0x306, 1),  # BREVE / COMBINING BREVE (Vrachy)
+    0xE7: (0x307, 1),  # SUPERIOR DOT / COMBINING DOT ABOVE
+    0xE8: (0x308, 1),  # UMLAUT, DIAERESIS / COMBINING DIAERESIS (Dialytika)
+    0xE9: (0x30C, 1),  # HACEK / COMBINING CARON
+    0xEA: (0x30A, 1),  # CIRCLE ABOVE, ANGSTROM / COMBINING RING ABOVE
+    0xEB: (0xFE20, 1),  # LIGATURE, FIRST HALF / COMBINING LIGATURE LEFT HALF
+    0xEC: (0xFE21, 1),  # LIGATURE, SECOND HALF / COMBINING LIGATURE RIGHT HALF
+    0xED: (0x315, 1),  # HIGH COMMA, OFF CENTER / COMBINING COMMA ABOVE RIGHT
+    0xEE: (0x30B, 1),  # DOUBLE ACUTE / COMBINING DOUBLE ACUTE ACCENT
+    0xEF: (0x310, 1),  # CANDRABINDU / COMBINING CANDRABINDU
+    0xF0: (0x327, 1),  # CEDILLA / COMBINING CEDILLA
+    0xF1: (0x328, 1),  # RIGHT HOOK, OGONEK / COMBINING OGONEK
+    0xF2: (0x323, 1),  # DOT BELOW / COMBINING DOT BELOW
+    0xF3: (0x324, 1),  # DOUBLE DOT BELOW / COMBINING DIAERESIS BELOW
+    0xF4: (0x325, 1),  # CIRCLE BELOW / COMBINING RING BELOW
+    0xF5: (0x333, 1),  # DOUBLE UNDERSCORE / COMBINING DOUBLE LOW LINE
+    0xF6: (0x332, 1),  # UNDERSCORE / COMBINING LOW LINE
+    0xF7: (0x326, 1),  # LEFT HOOK (COMMA BELOW) / COMBINING COMMA BELOW
+    0xF8: (0x31C, 1),  # RIGHT CEDILLA / COMBINING LEFT HALF RING BELOW
+    0xF9: (0x32E, 1),  # UPADHMANIYA / COMBINING BREVE BELOW
+    0xFA: (0xFE22, 1),  # DOUBLE TILDE, FIRST HALF / COMBINING DOUBLE TILDE LEFT HALF
+    0xFB: (0xFE23, 1),  # DOUBLE TILDE, SECOND HALF / COMBINING DOUBLE TILDE RIGHT HALF
+    0xFE: (0x313, 1),  # HIGH COMMA, CENTERED / COMBINING COMMA ABOVE (Psili)
 }
-CHARSET_67 = {  # Greek Symbols
-    0x61: (0x3B1, 0),  # GREEK SMALL LETTER ALPHA
-    0x62: (0x3B2, 0),  # GREEK SMALL LETTER BETA
-    0x63: (0x3B3, 0),  # GREEK SMALL LETTER GAMMA
-}
+
 CHARSET_4E = {  # Basic Cyrillic
     0x21: (0x21, 0),  # EXCLAMATION MARK
     0x22: (0x22, 0),  # QUOTATION MARK
@@ -16863,19 +16665,182 @@ CHARSET_4E = {  # Basic Cyrillic
     0x7D: (0x429, 0),  # CYRILLIC CAPITAL LETTER SHCHA
     0x7E: (0x427, 0),  # CYRILLIC CAPITAL LETTER CHE
 }
+
+CHARSET_51 = {  # Extended Cyrillic
+    0xC0: (0x491, 0),  # LOWERCASE GE WITH UPTURN / CYRILLIC SMALL LETTER GHE WITH UPTURN
+    0xC1: (0x452, 0),  # LOWERCASE DJE / CYRILLIC SMALL LETTER DJE (Serbian)
+    0xC2: (0x453, 0),  # CYRILLIC SMALL LETTER GJE
+    0xC3: (0x454, 0),  # LOWERCASE E / CYRILLIC SMALL LETTER UKRAINIAN IE
+    0xC4: (0x451, 0),  # CYRILLIC SMALL LETTER IO
+    0xC5: (0x455, 0),  # CYRILLIC SMALL LETTER DZE
+    0xC6: (0x456, 0),  # LOWERCASE I / CYRILLIC SMALL LETTER BYELORUSSIAN-UKRANIAN I
+    0xC7: (0x457, 0),  # LOWERCASE YI / CYRILLIC SMALL LETTER YI (Ukrainian)
+    0xC8: (0x458, 0),  # CYRILLIC SMALL LETTER JE
+    0xC9: (0x459, 0),  # CYRILLIC SMALL LETTER LJE
+    0xCA: (0x45A, 0),  # CYRILLIC SMALL LETTER NJE
+    0xCB: (0x45B, 0),  # LOWERCASE TSHE / CYRILLIC SMALL LETTER TSHE (Serbian)
+    0xCC: (0x45C, 0),  # CYRILLIC SMALL LETTER KJE
+    0xCD: (0x45E, 0),  # LOWERCASE SHORT U / CYRILLIC SMALL LETTER SHORT U (Byelorussian)
+    0xCE: (0x45F, 0),  # CYRILLIC SMALL LETTER DZHE
+    0xD0: (0x463, 0),  # CYRILLIC SMALL LETTER YAT
+    0xD1: (0x473, 0),  # CYRILLIC SMALL LETTER FITA
+    0xD2: (0x475, 0),  # CYRILLIC SMALL LETTER IZHITSA
+    0xD3: (0x46B, 0),  # CYRILLIC SMALL LETTER BIG YUS
+    0xDB: (0x5B, 0),  # OPENING SQUARE BRACKET / LEFT SQUARE BRACKET
+    0xDD: (0x5D, 0),  # CLOSING SQUARE BRACKET / RIGHT SQUARE BRACKET
+    0xDF: (0x5F, 0),  # SPACING UNDERSCORE / LOW LINE
+    0xE0: (0x490, 0),  # UPPERCASE GE WITH UPTURN / CYRILLIC CAPITAL LETTER GHE WITH UPTURN
+    0xE1: (0x402, 0),  # UPPERCASE DJE / CYRILLIC CAPITAL LETTER DJE (Serbian)
+    0xE2: (0x403, 0),  # CYRILLIC CAPITAL LETTER GJE
+    0xE3: (0x404, 0),  # UPPERCASE E / CYRILLIC CAPITAL LETTER UKRAINIAN IE
+    0xE4: (0x401, 0),  # CYRILLIC CAPITAL LETTER IO
+    0xE5: (0x405, 0),  # CYRILLIC CAPITAL LETTER DZE
+    0xE6: (0x406, 0),  # UPPERCASE I / CYRILLIC CAPITAL LETTER BYELORUSSIAN-UKRANIAN I
+    0xE7: (0x407, 0),  # UPPERCASE YI / CYRILLIC CAPITAL LETTER YI (Ukrainian)
+    0xE8: (0x408, 0),  # CYRILLIC CAPITAL LETTER JE
+    0xE9: (0x409, 0),  # CYRILLIC CAPITAL LETTER LJE
+    0xEA: (0x40A, 0),  # CYRILLIC CAPITAL LETTER NJE
+    0xEB: (0x40B, 0),  # UPPERCASE TSHE / CYRILLIC CAPITAL LETTER TSHE (Serbian)
+    0xEC: (0x40C, 0),  # CYRILLIC CAPITAL LETTER KJE
+    0xED: (0x40E, 0),  # UPPERCASE SHORT U / CYRILLIC CAPITAL LETTER SHORT U (Byelorussian)
+    0xEE: (0x40F, 0),  # CYRILLIC CAPITAL LETTER DZHE
+    0xEF: (0x42A, 0),  # CYRILLIC CAPITAL LETTER HARD SIGN
+    0xF0: (0x462, 0),  # CYRILLIC CAPITAL LETTER YAT
+    0xF1: (0x472, 0),  # CYRILLIC CAPITAL LETTER FITA
+    0xF2: (0x474, 0),  # CYRILLIC CAPITAL LETTER IZHITSA
+    0xF3: (0x46A, 0),  # CYRILLIC CAPITAL LETTER BIG YUS
+}
+
+CHARSET_53 = {  # Basic Greek
+    0x21: (0x300, 1),  # COMBINING GRAVE ACCENT
+    0x22: (0x301, 1),  # COMBINING ACUTE ACCENT
+    0x23: (0x308, 1),  # COMBINING DIAERESIS
+    0x24: (0x342, 1),  # COMBINING GREEK PERISPOMENI / CIRCUMFLEX
+    0x25: (0x313, 1),  # COMBINING COMMA ABOVE / SMOOTH BREATHING
+    0x26: (0x314, 1),  # COMBINING REVERSED COMMA ABOVE / ROUGH BREATHING
+    0x27: (0x345, 1),  # COMBINING GREEK YPOGEGRAMMENI / IOTA SUBSCRIPT
+    0x30: (0xAB, 0),  # LEFT-POINTING DOUBLE ANGLE QUOTATION MARK
+    0x31: (0xBB, 0),  # RIGHT-POINTING DOUBLE ANGLE QUOTATION MARK
+    0x32: (0x201C, 0),  # LEFT DOUBLE QUOTATION MARK
+    0x33: (0x201D, 0),  # RIGHT DOUBLE QUOTATION MARK
+    0x34: (0x374, 0),  # GREEK NUMERAL SIGN / UPPER PRIME
+    0x35: (0x375, 0),  # GREEK LOWER NUMERAL SIGN / LOWER PRIME
+    0x3B: (0x387, 0),  # GREEK ANO TELEIA / RAISED DOT, GREEK SEMICOLON
+    0x3F: (0x37E, 0),  # GREEK QUESTION MARK
+    0x41: (0x391, 0),  # GREEK CAPITAL LETTER ALPHA
+    0x42: (0x392, 0),  # GREEK CAPITAL LETTER BETA
+    0x44: (0x393, 0),  # GREEK CAPITAL LETTER GAMMA
+    0x45: (0x394, 0),  # GREEK CAPITAL LETTER DELTA
+    0x46: (0x395, 0),  # GREEK CAPITAL LETTER EPSILON
+    0x47: (0x3DA, 0),  # GREEK LETTER STIGMA
+    0x48: (0x3DC, 0),  # GREEK LETTER DIGAMMA
+    0x49: (0x396, 0),  # GREEK CAPITAL LETTER ZETA
+    0x4A: (0x397, 0),  # GREEK CAPITAL LETTER ETA
+    0x4B: (0x398, 0),  # GREEK CAPITAL LETTER THETA
+    0x4C: (0x399, 0),  # GREEK CAPITAL LETTER IOTA
+    0x4D: (0x39A, 0),  # GREEK CAPITAL LETTER KAPPA
+    0x4E: (0x39B, 0),  # GREEK CAPITAL LETTER LAMDA
+    0x4F: (0x39C, 0),  # GREEK CAPITAL LETTER MU
+    0x50: (0x39D, 0),  # GREEK CAPITAL LETTER NU
+    0x51: (0x39E, 0),  # GREEK CAPITAL LETTER XI
+    0x52: (0x39F, 0),  # GREEK CAPITAL LETTER OMICRON
+    0x53: (0x3A0, 0),  # GREEK CAPITAL LETTER PI
+    0x54: (0x3DE, 0),  # GREEK LETTER KOPPA
+    0x55: (0x3A1, 0),  # GREEK CAPITAL LETTER RHO
+    0x56: (0x3A3, 0),  # GREEK CAPITAL LETTER SIGMA
+    0x58: (0x3A4, 0),  # GREEK CAPITAL LETTER TAU
+    0x59: (0x3A5, 0),  # GREEK CAPITAL LETTER UPSILON
+    0x5A: (0x3A6, 0),  # GREEK CAPITAL LETTER PHI
+    0x5B: (0x3A7, 0),  # GREEK CAPITAL LETTER CHI
+    0x5C: (0x3A8, 0),  # GREEK CAPITAL LETTER PSI
+    0x5D: (0x3A9, 0),  # GREEK CAPITAL LETTER OMEGA
+    0x5E: (0x3E0, 0),  # GREEK LETTER SAMPI
+    0x61: (0x3B1, 0),  # GREEK SMALL LETTER ALPHA
+    0x62: (0x3B2, 0),  # GREEK SMALL LETTER BETA / SMALL LETTER BETA BEGINNING OF WORD
+    0x63: (0x3D0, 0),  # GREEK BETA SYMBOL / SMALL LETTER BETA MIDDLE OF WORD
+    0x64: (0x3B3, 0),  # GREEK SMALL LETTER GAMMA
+    0x65: (0x3B4, 0),  # GREEK SMALL LETTER DELTA
+    0x66: (0x3B5, 0),  # GREEK SMALL LETTER EPSILON
+    0x67: (0x3DB, 0),  # GREEK SMALL LETTER STIGMA
+    0x68: (0x3DD, 0),  # GREEK SMALL LETTER DIGAMMA
+    0x69: (0x3B6, 0),  # GREEK SMALL LETTER ZETA
+    0x6A: (0x3B7, 0),  # GREEK SMALL LETTER ETA
+    0x6B: (0x3B8, 0),  # GREEK SMALL LETTER THETA
+    0x6C: (0x3B9, 0),  # GREEK SMALL LETTER IOTA
+    0x6D: (0x3BA, 0),  # GREEK SMALL LETTER KAPPA
+    0x6E: (0x3BB, 0),  # GREEK SMALL LETTER LAMDA
+    0x6F: (0x3BC, 0),  # GREEK SMALL LETTER MU
+    0x70: (0x3BD, 0),  # GREEK SMALL LETTER NU
+    0x71: (0x3BE, 0),  # GREEK SMALL LETTER XI
+    0x72: (0x3BF, 0),  # GREEK SMALL LETTER OMICRON
+    0x73: (0x3C0, 0),  # GREEK SMALL LETTER PI
+    0x74: (0x3DF, 0),  # GREEK SMALL LETTER KOPPA
+    0x75: (0x3C1, 0),  # GREEK SMALL LETTER RHO
+    0x76: (0x3C3, 0),  # GREEK SMALL LETTER SIGMA
+    0x77: (0x3C2, 0),  # GREEK SMALL LETTER FINAL SIGMA / SMALL LETTER SIGMA END OF WORD
+    0x78: (0x3C4, 0),  # GREEK SMALL LETTER TAU
+    0x79: (0x3C5, 0),  # GREEK SMALL LETTER UPSILON
+    0x7A: (0x3C6, 0),  # GREEK SMALL LETTER PHI
+    0x7B: (0x3C7, 0),  # GREEK SMALL LETTER CHI
+    0x7C: (0x3C8, 0),  # GREEK SMALL LETTER PSI
+    0x7D: (0x3C9, 0),  # GREEK SMALL LETTER OMEGA
+    0x7E: (0x3E1, 0),  # GREEK SMALL LETTER SAMPI
+}
+
+CHARSET_62 = {  # Subscripts
+    0x28: (0x208D, 0),  # SUBSCRIPT OPENING PARENTHESIS / SUBSCRIPT LEFT PARENTHESIS
+    0x29: (0x208E, 0),  # SUBSCRIPT CLOSING PARENTHESIS / SUBSCRIPT RIGHT PARENTHESIS
+    0x2B: (0x208A, 0),  # SUBSCRIPT PLUS SIGN
+    0x2D: (0x208B, 0),  # SUBSCRIPT HYPHEN-MINUS / SUBSCRIPT MINUS
+    0x30: (0x2080, 0),  # SUBSCRIPT DIGIT ZERO
+    0x31: (0x2081, 0),  # SUBSCRIPT DIGIT ONE
+    0x32: (0x2082, 0),  # SUBSCRIPT DIGIT TWO
+    0x33: (0x2083, 0),  # SUBSCRIPT DIGIT THREE
+    0x34: (0x2084, 0),  # SUBSCRIPT DIGIT FOUR
+    0x35: (0x2085, 0),  # SUBSCRIPT DIGIT FIVE
+    0x36: (0x2086, 0),  # SUBSCRIPT DIGIT SIX
+    0x37: (0x2087, 0),  # SUBSCRIPT DIGIT SEVEN
+    0x38: (0x2088, 0),  # SUBSCRIPT DIGIT EIGHT
+    0x39: (0x2089, 0),  # SUBSCRIPT DIGIT NINE
+}
+
+CHARSET_67 = {  # Greek Symbols
+    0x61: (0x3B1, 0),  # GREEK SMALL LETTER ALPHA
+    0x62: (0x3B2, 0),  # GREEK SMALL LETTER BETA
+    0x63: (0x3B3, 0),  # GREEK SMALL LETTER GAMMA
+}
+
+CHARSET_70 = {  # Superscripts
+    0x28: (0x207D, 0),  # SUPERSCRIPT OPENING PARENTHESIS / SUPERSCRIPT LEFT PARENTHESIS
+    0x29: (0x207E, 0),  # SUPERSCRIPT CLOSING PARENTHESIS / SUPERSCRIPT RIGHT PARENTHESIS
+    0x2B: (0x207A, 0),  # SUPERSCRIPT PLUS SIGN
+    0x2D: (0x207B, 0),  # SUPERSCRIPT HYPHEN-MINUS / SUPERSCRIPT MINUS
+    0x30: (0x2070, 0),  # SUPERSCRIPT DIGIT ZERO
+    0x31: (0xB9, 0),  # SUPERSCRIPT DIGIT ONE
+    0x32: (0xB2, 0),  # SUPERSCRIPT DIGIT TWO
+    0x33: (0xB3, 0),  # SUPERSCRIPT DIGIT THREE
+    0x34: (0x2074, 0),  # SUPERSCRIPT DIGIT FOUR
+    0x35: (0x2075, 0),  # SUPERSCRIPT DIGIT FIVE
+    0x36: (0x2076, 0),  # SUPERSCRIPT DIGIT SIX
+    0x37: (0x2077, 0),  # SUPERSCRIPT DIGIT SEVEN
+    0x38: (0x2078, 0),  # SUPERSCRIPT DIGIT EIGHT
+    0x39: (0x2079, 0),  # SUPERSCRIPT DIGIT NINE
+}
+
+
 CODESETS = {
-    0x34: CHARSET_34,
-    0x45: CHARSET_45,
-    0x33: CHARSET_33,
-    0x32: CHARSET_32,
     0x31: CHARSET_31,
-    0x70: CHARSET_70,
+    0x32: CHARSET_32,
+    0x33: CHARSET_33,
+    0x34: CHARSET_34,
+    0x42: CHARSET_42,
+    0x45: CHARSET_45,
+    0x4E: CHARSET_4E,
     0x51: CHARSET_51,
     0x53: CHARSET_53,
-    0x42: CHARSET_42,
     0x62: CHARSET_62,
     0x67: CHARSET_67,
-    0x4E: CHARSET_4E,
+    0x70: CHARSET_70
 }
 
 # ODD_MAP for odd characters (all from III for now)
